requirements.txt
asgiref==3.5.0  
backports.zoneinfo==0.2.1  
Django==4.0.3  
django-cors-headers==3.11.0 
djangorestframework==3.13.1 
pip==20.0.2 
pkg-resources==0.0.0  
pytz==2021.3 
setuptools==44.0.0 
sqlparse==0.4.2